RemoteFS: report each share as a pool
Cinder RemoteFS based drivers reported total shares capacity as a
sum of all individual share capacities.

For this reason, a volume backend may be considered for placing
a volume, even if none of the shares used by the backend can actually
store the requested volume.

This patch changes this behavior, treating each share as a pool and
repoting the stats accordingly. This way, the scheduler can choose
a share for us.

The pool names are hashes of the share locations, ensuring that we
don't expose this kind of information to the end user.